+++ packages/gadget-sdk/src/testing/MockGadgetContextProvider.tsx
@@ -113,6 +113,7 @@ export function MockGadgetContextProvider({
         : {
             goBack: () => {},
             dismiss: () => {},
+            openSettings: async () => {},
             onExecute: () => {},
             onFooterChange: () => {},
             setDisplayQuery: () => {},